break observer out into separate module
